+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Multiline: added metric option "don't split line"
+- Multiline: added metric option "line by steps"
 
 ### Changed